refactor: regression helpers, evidence interface, logging, test split
•⁠  ⁠Extract _add_federated_output/_add_standalone_output helpers in
  acro_regression.py to eliminate repeated federated/standalone branching
•⁠  ⁠Clean up finalise_evidence() interface: accept evidence dict directly
  instead of relying on hidden _federated_evidence_store attribute
•⁠  ⁠Replace print() calls with logger.debug/warning in acro_stata_parser
  and ontology_handler; remove T201 global ignore from pyproject.toml
•⁠  ⁠Split test_initial.py into test_regressions.py, test_plots.py, and
  test_federated.py for better test organisation
•⁠  ⁠Fix pre-commit issues: codespell typo, stray print, inline import
