+++ acro/record.py
@@ -225,7 +225,6 @@ class Records:
         self.blocked_extensions: list[str] = [
             ext.lower() for ext in (blocked_extensions or [])
         ]
-        self._federated_evidence_store: dict = {}
 
     def add(
         self,
@@ -451,7 +450,7 @@ class Records:
         outputs: str = ""
         for _, record in self.results.items():
             outputs += str(record) + "\n"
-        print(outputs)
+        print(outputs)  # noqa: T201
         return outputs
 
     def validate_outputs(self) -> None:
@@ -581,8 +580,8 @@ class Records:
                     start = 1 + writer.sheets[output_id].max_row
                     table.to_excel(writer, sheet_name=output_id, startrow=start)
 
-    def finalise_evidence(self, path: str) -> dict:
-        """Serialise federated evidence to CSV files and return the manifest dict.  # noqa: D212,D213,D413.
+    def finalise_evidence(self, path: str, evidence_store: dict | None = None) -> dict:
+        """Serialise federated evidence to CSV files and return the manifest dict.
 
         Each interim table (DataFrame) is saved as a separate CSV file in *path*.
         The returned dictionary is suitable for writing to ``evidence.json``.
@@ -591,6 +590,10 @@ class Records:
         ----------
         path : str
             Directory where CSV files and ``evidence.json`` will be written.
+        evidence_store : dict, optional
+            The evidence dictionary to serialise.  When ``None`` an empty dict
+            is used, producing an empty manifest.  Callers should pass
+            ``getattr(self_acro, "_federated_evidence", {})``.
 
         Returns
         -------
@@ -600,13 +603,8 @@ class Records:
         os.makedirs(path, exist_ok=True)
         outputs: dict[str, Any] = {}
 
-        # Merge evidence from both Tables and Regression (stored on the ACRO obj
-        # via _federated_evidence on each mixin).  At this point self is a Records
-        # instance; the caller (ACRO.finalise) passes the merged dict directly.
-        # We receive it as a parameter so the method stays on Records.
-        # NOTE: the actual evidence dict is passed in via the ``_evidence`` attr
-        # attached by ACRO.finalise() just before calling this method.
-        evidence_store: dict = getattr(self, "_federated_evidence_store", {})
+        if evidence_store is None:
+            evidence_store = {}
 
         for uid, entry in evidence_store.items():
             table_files: dict[str, str] = {}
