+++ test/test_initial.py
@@ -3,7 +3,6 @@
 import json
 import logging
 import os
-import pathlib
 import shutil
 import tempfile
 
@@ -17,7 +16,6 @@ import numpy as np
 import pandas as pd
 import pytest
 import rdflib
-import statsmodels.api as sm
 
 from acro import (
     ACRO,
@@ -27,7 +25,6 @@ from acro import (
     table_utils,
     utils,
 )
-from acro.acro_tables import _rounded_survival_table  # , crosstab_with_totals
 from acro.aggregationfunctions import agg_nk, agg_p_percent, agg_threshold
 from acro.ontology_handler import (
     PREFIX,
@@ -307,88 +304,6 @@ def test_pivot_table_with_aggfunc_sum(data, acro):
     output_1 = results.get_index(1)
     assert output_0.status in ("review", "fail", "pass")
     assert output_1.status in ("review", "fail", "pass")
-    shutil.rmtree(PATH)
-
-
-def test_ols(data, acro):
-    """Ordinary Least Squares test."""
-    new_df = data[["inc_activity", "inc_grants", "inc_donations", "total_costs"]]
-    new_df = new_df.dropna()
-    # OLS too few Dof
-    endog = new_df.inc_activity.iloc[0:10]
-    exog = new_df[["inc_grants", "inc_donations", "total_costs"]].iloc[0:10]
-    exog = add_constant(exog)
-    results = acro.ols(endog, exog)
-    assert results.df_resid == 6
-    res = acro.results.get_index(-1)
-    assert res.status == "fail"
-    acro.remove_output(res.uid)
-
-    # OLS
-    endog = new_df.inc_activity
-    exog = new_df[["inc_grants", "inc_donations", "total_costs"]]
-    exog = add_constant(exog)
-    results = acro.ols(endog, exog)
-    assert results.df_resid == 807
-    assert results.rsquared == pytest.approx(0.894, 0.001)
-    # OLSR
-    results = acro.olsr(
-        formula="inc_activity ~ inc_grants + inc_donations + total_costs", data=new_df
-    )
-    assert results.df_resid == 807
-    assert results.rsquared == pytest.approx(0.894, 0.001)
-    # Finalise
-    results = acro.finalise(PATH)
-    output_0 = results.get_index(0)
-    output_1 = results.get_index(1)
-    assert output_0.status == "pass"
-    assert output_1.status == "pass"
-    shutil.rmtree(PATH)
-
-
-def test_probit_logit(data, acro):
-    """Probit and Logit tests."""
-    new_df = data[
-        ["survivor", "inc_activity", "inc_grants", "inc_donations", "total_costs"]
-    ]
-    new_df = new_df.dropna()
-    endog = new_df["survivor"].astype("category").cat.codes  # numeric
-    endog.name = "survivor"
-    exog = new_df[["inc_activity", "inc_grants", "inc_donations", "total_costs"]]
-    exog = add_constant(exog)
-    # Probit
-    results = acro.probit(endog, exog)
-    assert results.df_resid == 806
-    assert results.prsquared == pytest.approx(0.208, 0.01)
-    # Logit
-    results = acro.logit(endog, exog)
-    assert results.df_resid == 806
-    assert results.prsquared == pytest.approx(0.214, 0.01)
-    # ProbitR
-    new_df["survivor"] = new_df["survivor"].astype("category").cat.codes
-    results = acro.probitr(
-        formula="survivor ~ inc_activity + inc_grants + inc_donations + total_costs",
-        data=new_df,
-    )
-    assert results.df_resid == 806
-    assert results.prsquared == pytest.approx(0.208, 0.01)
-    # LogitR
-    results = acro.logitr(
-        formula="survivor ~ inc_activity + inc_grants + inc_donations + total_costs",
-        data=new_df,
-    )
-    assert results.df_resid == 806
-    assert results.prsquared == pytest.approx(0.214, 0.01)
-    # Finalise
-    results = acro.finalise(PATH)
-    output_0 = results.get_index(0)
-    output_1 = results.get_index(1)
-    output_2 = results.get_index(2)
-    output_3 = results.get_index(3)
-    assert output_0.status == "pass"
-    assert output_1.status == "pass"
-    assert output_2.status == "pass"
-    assert output_3.status == "pass"
     shutil.rmtree(PATH)
 
 
@@ -694,59 +609,6 @@ def test_single_values_column(data, acro):
         _ = acro.crosstab(data.year, data.grant_type, values=None, aggfunc="mean")
 
 
-def test_surv_func(acro):
-    """Test survival tables and plots."""
-    try:
-        data = sm.datasets.get_rdataset("flchain", "survival").data
-    except Exception:
-        np.random.seed(42)
-        mock_data = pd.DataFrame(
-            {
-                "futime": np.random.exponential(100, 500),
-                "death": np.random.binomial(1, 0.3, 500),
-                "sex": np.random.choice(["F", "M"], 500),
-            }
-        )
-        data = mock_data
-
-    data = data.loc[data.sex == "F", :]
-    _ = acro.surv_func(data.futime, data.death, output="table")
-    output = acro.results.get_index(0)
-    assert output.status in ["fail", "review"]
-    assert "KaplanMeier" in output.summary
-
-    filename = os.path.normpath("acro_artifacts/kaplan-meier_0.png")
-    _ = acro.surv_func(data.futime, data.death, output="plot")
-    assert os.path.exists(filename)
-    acro.add_exception("output_0", "I need this")
-    acro.add_exception("output_1", "Let me have it")
-
-    foo = acro.surv_func(data.futime, data.death, output="something_else")
-    assert foo is None
-
-    results: Records = acro.finalise(path=PATH)
-    output_1 = results.get_index(1)
-    assert output_1.output == [filename]
-    shutil.rmtree(PATH)
-
-
-def test_rounded_survival_table():
-    """Test the rounded_survival_table function for survival analysis."""
-    survival_table = pd.DataFrame(
-        {
-            "Surv prob": [1.0, 0.95, 0.90, 0.85, 0.80],
-            "num at risk": [100, 95, 85, 75, 60],
-            "num events": [0, 5, 10, 10, 15],
-        }
-    )
-    result = _rounded_survival_table(survival_table.copy())
-    assert "rounded_survival_fun" in result.columns
-    assert len(result) == 5
-    assert all(
-        (result["rounded_survival_fun"] >= 0) & (result["rounded_survival_fun"] <= 1)
-    )
-
-
 def test_zeros_are_not_disclosive(data, acro):
     """Test that zeros are handled as not disclosive when `zeros_are_disclosive=False`."""
     acro.sdc_checks.risk_appetite["zeros_are_disclosive"] = False
@@ -1027,81 +889,6 @@ def test_crosstab_with_manual_totals_with_suppression_with_two_aggfunc(data, acr
     assert acro.results.get_index(0).status in {"review", "fail"}
 
 
-def test_histogram_disclosive(acro, caplog):
-    """Test a disclosive histogram under the new suppression workflow."""
-    small_data = pd.DataFrame({"value": [1, 2, 3]})
-    result = acro.hist(small_data, "value")
-
-    assert result == ""
-    acro.add_exception("output_0", "Let me have it")
-    results: Records = acro.finalise(path=PATH)
-    output_0 = results.get_index(0)
-
-    assert (
-        "Histogram will not be shown as the value column is disclosive." in caplog.text
-    )
-    assert output_0.status == "fail"
-    assert output_0.output == []
-    shutil.rmtree(PATH)
-
-
-def test_histogram_non_disclosive(acro):
-    """Test a non-disclosive histogram with a larger synthetic dataset."""
-    rng = np.random.default_rng(42)
-    data = pd.DataFrame({"value": rng.normal(size=2000)})
-
-    result = acro.hist(data, "value", bins=5)
-
-    assert result is not None
-    assert os.path.exists(result)
-    assert os.path.normpath(result) == os.path.normpath(result)
-    acro.add_exception("output_0", "Let me have it")
-    results: Records = acro.finalise(path=PATH)
-    output_0 = results.get_index(0)
-    assert output_0.output == [os.path.normpath(result)]
-    assert output_0.status == "review"
-    shutil.rmtree(PATH)
-
-
-def test_pie_disclosive(acro, caplog):
-    """Test a disclosive pie chart (a category has fewer than threshold observations)."""
-    shutil.rmtree("acro_artifacts", ignore_errors=True)
-    shutil.rmtree(PATH, ignore_errors=True)
-
-    df = pd.DataFrame(
-        {"grant_type": (["A"] * 20) + (["B"] * 15) + (["C"] * 12) + (["D"] * 5)}
-    )
-
-    _ = acro.pie(df, "grant_type", filename="pie.png")
-    # When disclosive and suppressed, file should NOT be created
-    acro.add_exception("output_0", "Let me have it")
-    results: Records = acro.finalise(path=PATH)
-    output_0 = results.get_index(0)
-
-    assert (
-        "Pie chart will not be shown as the grant_type column is disclosive."
-        in caplog.text
-    )
-    assert output_0.status == "fail"
-    shutil.rmtree(PATH)
-
-
-def test_pie_non_disclosive(data, acro):
-    """Test a non-disclosive pie chart (all categories meet the threshold)."""
-    shutil.rmtree("acro_artifacts", ignore_errors=True)
-    shutil.rmtree(PATH, ignore_errors=True)
-    filename = os.path.normpath("acro_artifacts/pie_0.png")
-    result = acro.pie(data, "grant_type", filename="pie.png")
-    assert os.path.normpath(result) == filename
-    assert os.path.exists(filename)
-    acro.add_exception("output_0", "Let me have it")
-    results: Records = acro.finalise(path=PATH)
-    output_0 = results.get_index(0)
-    assert output_0.output == [filename]
-    assert output_0.status == "review"
-    shutil.rmtree(PATH)
-
-
 def test_finalise_with_existing_path(data, acro, caplog):
     """Test using a path that already exists when finalising."""
     _ = acro.crosstab(data.year, data.grant_type)
@@ -1188,7 +975,6 @@ def test_finalise_interactive(data):
     assert read0.exception == "Oh, please..."
     assert orig1.exception == "Suppression automatically applied where needed"
     assert read1.exception == "Suppression automatically applied where needed"
-    print(orig0.exception)
     # check SDC outcome DataFrame
     orig_df = orig0.output[0].reset_index()
     read_df = read0.output[0]
@@ -1265,226 +1051,6 @@ def test_crosstab_multi_aggfunc(data):
     )
     assert isinstance(table2, pd.DataFrame)
     assert table2.columns.nlevels == 2
-
-
-def acro_federated() -> ACRO:
-    """Initialise ACRO in federated mode."""
-    return ACRO(federated=True)
-
-
-def test_federated_flag_set_via_constructor():
-    """ACRO(federated=True) should set the federated attribute."""
-    acro = ACRO(federated=True)
-    assert acro.federated is True
-
-
-def test_federated_flag_default_is_false():
-    """ACRO() should default to federated=False."""
-    acro = ACRO()
-    assert acro.federated is False
-
-
-def test_federated_flag_from_yaml(tmp_path):
-    """Federated: true in yaml should set federated=True when no constructor override."""
-    yaml_content = (
-        "safe_threshold: 10\n"
-        "safe_dof_threshold: 10\n"
-        "safe_nk_n: 2\n"
-        "safe_nk_k: 0.90\n"
-        "safe_pratio_p: 0.10\n"
-        "check_missing_values: false\n"
-        "survival_safe_threshold: 10\n"
-        "zeros_are_disclosive: true\n"
-        "federated: true\n"
-    )
-    cfg_file = tmp_path / "fed_test.yaml"
-    cfg_file.write_text(yaml_content)
-
-    acro_pkg = pathlib.Path(__file__).parent.parent / "acro"
-    cfg_dest = acro_pkg / "fed_test.yaml"
-    cfg_dest.write_text(yaml_content)
-    try:
-        acro = ACRO(config="fed_test")
-        assert acro.federated is True
-    finally:
-        cfg_dest.unlink(missing_ok=True)
-
-
-def test_federated_constructor_overrides_yaml():
-    """Constructor federated=False should override yaml federated: true."""
-    yaml_content = (
-        "safe_threshold: 10\n"
-        "safe_dof_threshold: 10\n"
-        "safe_nk_n: 2\n"
-        "safe_nk_k: 0.90\n"
-        "safe_pratio_p: 0.10\n"
-        "check_missing_values: false\n"
-        "survival_safe_threshold: 10\n"
-        "zeros_are_disclosive: true\n"
-        "federated: true\n"
-    )
-    acro_pkg = pathlib.Path(__file__).parent.parent / "acro"
-    cfg_dest = acro_pkg / "fed_override_test.yaml"
-    cfg_dest.write_text(yaml_content)
-    try:
-        acro = ACRO(config="fed_override_test", federated=False)
-        assert acro.federated is False
-    finally:
-        cfg_dest.unlink(missing_ok=True)
-
-
-def test_federated_crosstab_skips_checks_and_stores_evidence(data):
-    """In federated mode, crosstab should store evidence and skip checks."""
-    acro = ACRO(federated=True)
-    table = acro.crosstab(data.year, data.grant_type)
-
-    # The returned table should still be a valid DataFrame
-    assert isinstance(table, pd.DataFrame)
-    assert not table.empty
-
-    # No records should have been written to results (checks were skipped)
-    assert acro.results.output_id == 1
-    assert len(acro.results.results) == 0
-
-    # Evidence should have been accumulated
-    assert "output_0" in acro._federated_evidence
-    entry = acro._federated_evidence["output_0"]
-    assert "count_table" in entry["interim_tables"]
-    assert entry["analysis_names"] == ["FrequencyTable"]
-
-
-def test_federated_pivot_table_stores_evidence(data):
-    """In federated mode, pivot_table should store evidence."""
-    acro = ACRO(federated=True)
-    _ = acro.pivot_table(
-        data, index=["grant_type"], values=["inc_grants"], aggfunc=["mean"]
-    )
-    assert "output_0" in acro._federated_evidence
-    entry = acro._federated_evidence["output_0"]
-    assert entry["analysis_names"] == ["Mean"]
-
-
-def test_federated_ols_stores_evidence(data):
-    """In federated mode, ols should store evidence (DoF) and skip checks."""
-    acro = ACRO(federated=True)
-    new_df = data[
-        ["inc_activity", "inc_grants", "inc_donations", "total_costs"]
-    ].dropna()
-    endog = new_df.inc_activity
-    exog = new_df[["inc_grants", "inc_donations", "total_costs"]]
-    exog = add_constant(exog)
-    results = acro.ols(endog, exog)
-
-    # statsmodels result should still be returned
-    assert results.df_resid == 807
-
-    # No record written
-    assert len(acro.results.results) == 0
-
-    # Evidence stored
-    assert "output_0" in acro._federated_evidence
-    entry = acro._federated_evidence["output_0"]
-    assert entry["analysis_names"] == ["GeneralLinearModel"]
-    assert entry["dof"] == 807
-
-
-def test_federated_finalise_writes_evidence_json(data):
-    """Finalise() in federated mode should produce evidence.json and CSV files."""
-    acro = ACRO(federated=True)
-    _ = acro.crosstab(data.year, data.grant_type)
-    result = acro.finalise(PATH)
-
-    assert result is not None
-
-    # evidence.json must exist
-    evidence_path = os.path.normpath(f"{PATH}/evidence.json")
-    assert os.path.exists(evidence_path), "evidence.json not found"
-
-    with open(evidence_path, encoding="utf-8") as fh:
-        evidence = json.load(fh)
-
-    assert "version" in evidence
-    assert "outputs" in evidence
-    assert "output_0" in evidence["outputs"]
-
-    entry = evidence["outputs"]["output_0"]
-    assert "analysis_names" in entry
-    assert "interim_tables" in entry
-
-    # At least the count_table CSV should exist
-    count_table_file = entry["interim_tables"].get("count_table")
-    assert count_table_file is not None
-    assert os.path.exists(os.path.normpath(f"{PATH}/{count_table_file}"))
-
-    # config.json should also be present
-    assert os.path.exists(os.path.normpath(f"{PATH}/config.json"))
-
-    # results.json should NOT exist in federated mode
-    assert not os.path.exists(os.path.normpath(f"{PATH}/results.json"))
-
-    shutil.rmtree(PATH)
-
-
-def test_federated_finalise_multiple_outputs(data):
-    """Federated finalise with multiple outputs produces one entry per output."""
-    acro = ACRO(federated=True)
-    _ = acro.crosstab(data.year, data.grant_type)
-    _ = acro.pivot_table(
-        data, index=["grant_type"], values=["inc_grants"], aggfunc=["mean"]
-    )
-
-    result = acro.finalise(PATH)
-    assert result is not None
-
-    with open(os.path.normpath(f"{PATH}/evidence.json"), encoding="utf-8") as fh:
-        evidence = json.load(fh)
-
-    assert len(evidence["outputs"]) == 2
-    assert "output_0" in evidence["outputs"]
-    assert "output_1" in evidence["outputs"]
-
-    shutil.rmtree(PATH)
-
-
-def test_federated_finalise_regression(data):
-    """Federated finalise for a regression writes DoF evidence."""
-    acro = ACRO(federated=True)
-    new_df = data[
-        ["inc_activity", "inc_grants", "inc_donations", "total_costs"]
-    ].dropna()
-    endog = new_df.inc_activity
-    exog = new_df[["inc_grants", "inc_donations", "total_costs"]]
-    exog = add_constant(exog)
-    _ = acro.ols(endog, exog)
-
-    result = acro.finalise(PATH)
-    assert result is not None
-
-    with open(os.path.normpath(f"{PATH}/evidence.json"), encoding="utf-8") as fh:
-        evidence = json.load(fh)
-
-    entry = evidence["outputs"]["output_0"]
-    assert entry["dof"] == 807
-    assert entry["analysis_names"] == ["GeneralLinearModel"]
-    # No CSV interim tables for regression (DoF is a scalar)
-    assert entry["interim_tables"] == {}
-
-    shutil.rmtree(PATH)
-
-
-def test_local_mode_unaffected_by_federated_flag(data):
-    """Non-federated ACRO should still produce results.json as before."""
-    acro = ACRO(suppress=False)
-    assert acro.federated is False
-    _ = acro.crosstab(data.year, data.grant_type)
-    acro.add_exception("output_0", "Let me have it")
-    result = acro.finalise(PATH)
-
-    assert result is not None
-    assert os.path.exists(os.path.normpath(f"{PATH}/results.json"))
-    assert not os.path.exists(os.path.normpath(f"{PATH}/evidence.json"))
-
-    shutil.rmtree(PATH)
 
 
 def test_get_catdtype_string_series():
@@ -1871,287 +1437,6 @@ def test_record_table_output_round_mitigation(data):
     assert "Rounding" in output.exception
 
 
-def test_store_federated_evidence_with_dataframe_dof():
-    """
-    _store_federated_evidence serialises DataFrame DoF to CSV string.
-
-    KaplanMeier's DoF comes from TableModelDetails.get_count_table()-1,
-    which is a DataFrame — this hits the isinstance(dof, DataFrame) branch.
-    """
-    acro_obj = ACRO(federated=True)
-    np.random.seed(42)
-    mock_data = pd.DataFrame(
-        {
-            "futime": np.random.exponential(100, 300),
-            "death": np.random.binomial(1, 0.3, 300),
-            "sex": np.random.choice(["F", "M"], 300),
-        }
-    )
-    mock_data = mock_data.loc[mock_data.sex == "F"]
-    _ = acro_obj.surv_func(mock_data.futime, mock_data.death, output="table")
-    entry = acro_obj._federated_evidence.get("output_0", {})
-    dof_val = entry.get("dof")
-    assert dof_val is not None
-    assert isinstance(dof_val, str)
-    assert "\n" in dof_val
-
-
-def test_surv_func_federated_table():
-    """Surv_func in federated mode with output='table' returns survival table."""
-    acro_obj = ACRO(federated=True)
-    np.random.seed(42)
-    mock_data = pd.DataFrame(
-        {
-            "futime": np.random.exponential(100, 500),
-            "death": np.random.binomial(1, 0.3, 500),
-        }
-    )
-    result = acro_obj.surv_func(mock_data.futime, mock_data.death, output="table")
-    assert isinstance(result, pd.DataFrame)
-    assert acro_obj.results.output_id == 1
-    assert len(acro_obj.results.results) == 0
-
-
-def test_surv_func_federated_plot_blocked_extension():
-    """Surv_func federated with blocked extension returns None (line 473)."""
-    acro_obj = ACRO(federated=True)
-    acro_obj.results.blocked_extensions = [".png"]
-    np.random.seed(42)
-    mock_data = pd.DataFrame(
-        {
-            "futime": np.random.exponential(100, 500),
-            "death": np.random.binomial(1, 0.3, 500),
-        }
-    )
-    result = acro_obj.surv_func(
-        mock_data.futime, mock_data.death, output="plot", filename="km.png"
-    )
-    assert result is None
-
-
-def test_surv_func_federated_returns_none_for_unknown_output():
-    """Surv_func federated with unknown output type returns None."""
-    acro_obj = ACRO(federated=True)
-    np.random.seed(42)
-    mock_data = pd.DataFrame(
-        {
-            "futime": np.random.exponential(100, 500),
-            "death": np.random.binomial(1, 0.3, 500),
-        }
-    )
-    result = acro_obj.surv_func(
-        mock_data.futime, mock_data.death, output="something_unknown"
-    )
-    assert result is None
-
-
-def test_surv_func_returns_none_for_unknown_output():
-    """Surv_func local mode with unknown output type returns None."""
-    acro_obj = ACRO(suppress=True)
-    np.random.seed(42)
-    mock_data = pd.DataFrame(
-        {
-            "futime": np.random.exponential(100, 500),
-            "death": np.random.binomial(1, 0.3, 500),
-        }
-    )
-    result = acro_obj.surv_func(mock_data.futime, mock_data.death, output="invalid")
-    assert result is None
-
-
-def test_surv_func_suppress_plot_blocked_extension():
-    """Surv_func with suppress=True, blocked extension returns None (line 734)."""
-    acro_obj = ACRO(suppress=True)
-    acro_obj.results.blocked_extensions = [".png"]
-    np.random.seed(7)
-    mock_data = pd.DataFrame(
-        {
-            "futime": np.random.exponential(100, 500),
-            "death": np.random.binomial(1, 0.3, 500),
-        }
-    )
-    result = acro_obj.surv_func(
-        mock_data.futime, mock_data.death, output="plot", filename="km.png"
-    )
-    assert result is None
-
-
-def test_hist_federated_returns_none():
-    """Hist() in federated mode stores evidence and returns None."""
-    acro_obj = ACRO(federated=True)
-    rng = np.random.default_rng(0)
-    df = pd.DataFrame({"val": rng.normal(size=500)})
-    result = acro_obj.hist(df, "val")
-    assert result is None
-    assert "output_0" in acro_obj._federated_evidence
-    assert len(acro_obj.results.results) == 0
-
-
-def test_hist_blocked_extension():
-    """Hist() with a blocked extension returns None (line 856)."""
-    acro_obj = ACRO(suppress=True)
-    acro_obj.results.blocked_extensions = [".png"]
-    rng = np.random.default_rng(1)
-    df = pd.DataFrame({"val": rng.normal(size=500)})
-    result = acro_obj.hist(df, "val", filename="histogram.png")
-    assert result is None
-    # No output should have been recorded
-    assert len(acro_obj.results.results) == 0
-
-
-def test_hist_non_disclosive_no_suppress(data):
-    """Hist() with suppress=False always records output — file path is saved even if disclosive."""
-    acro_obj = ACRO(suppress=False)
-    result = acro_obj.hist(data, "inc_grants", bins=10)
-    output = acro_obj.results.get_index(0)
-    assert output.output_type == "histogram"
-    # With suppress=False the histogram is always saved and path is recorded
-    assert result is not None
-    assert result != ""
-    assert output.output == [os.path.normpath(result)]
-
-
-def test_pie_federated_returns_none(data):
-    """Pie() in federated mode stores evidence and returns None (line 1032)."""
-    acro_obj = ACRO(federated=True)
-    result = acro_obj.pie(data, "grant_type")
-    assert result is None
-    assert "output_0" in acro_obj._federated_evidence
-    assert len(acro_obj.results.results) == 0
-
-
-def test_pie_blocked_extension(data):
-    """Pie() with a blocked extension returns None."""
-    acro_obj = ACRO(suppress=True)
-    acro_obj.results.blocked_extensions = [".png"]
-    result = acro_obj.pie(data, "grant_type", filename="pie.png")
-    assert result is None
-    assert len(acro_obj.results.results) == 0
-
-
-def test_pie_records_output_non_disclosive(data):
-    """Pie() on non-disclosive data records output with correct type and non-fail status."""
-    acro_obj = ACRO(suppress=False)
-    acro_obj.pie(data, "grant_type")
-    output = acro_obj.results.get_index(0)
-    assert output.output_type == "pie chart"
-    assert output.status in ("pass", "review")
-
-
-def test_olsr_federated(data):
-    """Olsr() in federated mode stores evidence and skips checks."""
-    acro_obj = ACRO(federated=True)
-    new_df = data[
-        ["inc_activity", "inc_grants", "inc_donations", "total_costs"]
-    ].dropna()
-    results = acro_obj.olsr(
-        formula="inc_activity ~ inc_grants + inc_donations + total_costs", data=new_df
-    )
-    assert results.df_resid == 807
-    assert len(acro_obj.results.results) == 0
-    assert "output_0" in acro_obj._federated_evidence
-    entry = acro_obj._federated_evidence["output_0"]
-    assert entry["analysis_names"] == ["GeneralLinearModel"]
-
-
-def test_logitr_federated(data):
-    """Logitr() in federated mode stores evidence and skips checks."""
-    acro_obj = ACRO(federated=True)
-    new_df = data[
-        ["survivor", "inc_activity", "inc_grants", "inc_donations", "total_costs"]
-    ].dropna()
-    new_df = new_df.copy()
-    new_df["survivor"] = new_df["survivor"].astype("category").cat.codes
-    results = acro_obj.logitr(
-        formula="survivor ~ inc_activity + inc_grants + inc_donations + total_costs",
-        data=new_df,
-    )
-    assert results.df_resid == 806
-    assert len(acro_obj.results.results) == 0
-    assert "output_0" in acro_obj._federated_evidence
-    entry = acro_obj._federated_evidence["output_0"]
-    assert entry["analysis_names"] == ["Logit"]
-
-
-def test_probitr_federated(data):
-    """Probitr() in federated mode stores evidence and skips checks."""
-    acro_obj = ACRO(federated=True)
-    new_df = data[
-        ["survivor", "inc_activity", "inc_grants", "inc_donations", "total_costs"]
-    ].dropna()
-    new_df = new_df.copy()
-    new_df["survivor"] = new_df["survivor"].astype("category").cat.codes
-    results = acro_obj.probitr(
-        formula="survivor ~ inc_activity + inc_grants + inc_donations + total_costs",
-        data=new_df,
-    )
-    assert results.df_resid == 806
-    assert len(acro_obj.results.results) == 0
-    assert "output_0" in acro_obj._federated_evidence
-    entry = acro_obj._federated_evidence["output_0"]
-    assert entry["analysis_names"] == ["Probit"]
-
-
-def test_probit_federated(data):
-    """Probit() in federated mode stores evidence and skips checks."""
-    acro_obj = ACRO(federated=True)
-    new_df = data[
-        ["survivor", "inc_activity", "inc_grants", "inc_donations", "total_costs"]
-    ].dropna()
-    endog = new_df["survivor"].astype("category").cat.codes
-    endog.name = "survivor"
-    exog = new_df[["inc_activity", "inc_grants", "inc_donations", "total_costs"]]
-    exog = add_constant(exog)
-    results = acro_obj.probit(endog, exog)
-    assert results.df_resid == 806
-    assert len(acro_obj.results.results) == 0
-    assert "output_0" in acro_obj._federated_evidence
-    entry = acro_obj._federated_evidence["output_0"]
-    assert entry["analysis_names"] == ["Probit"]
-
-
-def test_logit_federated(data):
-    """Logit() in federated mode stores evidence and skips checks."""
-    acro_obj = ACRO(federated=True)
-    new_df = data[
-        ["survivor", "inc_activity", "inc_grants", "inc_donations", "total_costs"]
-    ].dropna()
-    endog = new_df["survivor"].astype("category").cat.codes
-    endog.name = "survivor"
-    exog = new_df[["inc_activity", "inc_grants", "inc_donations", "total_costs"]]
-    exog = add_constant(exog)
-    results = acro_obj.logit(endog, exog)
-    assert results.df_resid == 806
-    assert len(acro_obj.results.results) == 0
-    assert "output_0" in acro_obj._federated_evidence
-    entry = acro_obj._federated_evidence["output_0"]
-    assert entry["analysis_names"] == ["Logit"]
-
-
-def test_show_fair_summaries_dict_and_scalar(data):
-    """Show_fair_summaries() covers the dict and scalar branches."""
-    acro_obj = ACRO(suppress=False)
-    _ = acro_obj.crosstab(data.year, data.grant_type)
-    result = acro_obj.show_fair_summaries()
-    assert isinstance(result, str)
-    assert "output_0" in result
-
-
-def test_show_fair_summaries_regression(data):
-    """Show_fair_summaries() works when fair dict has nested dict values."""
-    acro_obj = ACRO(suppress=False)
-    new_df = data[
-        ["inc_activity", "inc_grants", "inc_donations", "total_costs"]
-    ].dropna()
-    endog = new_df.inc_activity
-    exog = new_df[["inc_grants", "inc_donations", "total_costs"]]
-    exog = add_constant(exog)
-    _ = acro_obj.ols(endog, exog)
-    result = acro_obj.show_fair_summaries()
-    assert isinstance(result, str)
-    assert len(result) > 0
-
-
 def test_add_to_acro_function(data, monkeypatch):
     """Add_to_acro() exercises by scanning a directory and creating results."""
     src_path = "test_add_to_acro"
@@ -2197,9 +1482,8 @@ def test_records_add_with_none_defaults():
 def test_finalise_evidence_dof_as_csv_string():
     """Finalise_evidence() writes a CSV file when dof is a multiline string."""
     records = Records()
-    # Simulate the evidence store the way ACRO.finalise() sets it
     dof_csv = "idx,val\n0,10\n1,20\n"  # multiline CSV string
-    records._federated_evidence_store = {
+    evidence_store = {
         "output_0": {
             "command": "test",
             "analysis_names": ["FrequencyTable"],
@@ -2209,7 +1493,7 @@ def test_finalise_evidence_dof_as_csv_string():
         }
     }
     with tempfile.TemporaryDirectory() as tmp:
-        manifest = records.finalise_evidence(tmp)
+        manifest = records.finalise_evidence(tmp, evidence_store)
         # dof_file should have been written
         entry = manifest["outputs"]["output_0"]
         dof_file = entry["dof"]
@@ -2312,32 +1596,6 @@ def test_append_rounded_margins_unsupported_aggfunc():
     df = pd.DataFrame({"a": [10, 20], "b": [30, 40]}, index=[1, 2])
     result = table_utils.append_rounded_margins(df, "max", "All", 5)
     pd.testing.assert_frame_equal(result, df)
-
-
-def test_surv_func_suppress_true_table_records_exception():
-    """Surv_func with suppress=True and output='table' adds exception."""
-    acro_obj = ACRO(suppress=True)
-    np.random.seed(99)
-    mock_data = pd.DataFrame(
-        {
-            "futime": np.random.exponential(100, 500),
-            "death": np.random.binomial(1, 0.3, 500),
-        }
-    )
-    result = acro_obj.surv_func(mock_data.futime, mock_data.death, output="table")
-    assert isinstance(result, pd.DataFrame)
-    output = acro_obj.results.get_index(0)
-    assert output.status == "review"
-    assert "Events Reported" in output.exception
-
-
-def test_hist_federated_no_output_recorded(data):
-    """Hist() federated gate: no result is added to results."""
-    acro_obj = ACRO(federated=True)
-    result = acro_obj.hist(data, "inc_grants", bins=5)
-    assert result is None
-    assert acro_obj.results.output_id == 1
-    assert len(acro_obj.results.results) == 0
 
 
 def test_agg_p_percent_normal_violation():
@@ -2503,13 +1761,15 @@ def test_is_uri_string_is_not_uri():
     assert is_uri("http://example.org") is False
 
 
-def test_print_nested_dict_prints_without_error(capsys):
-    """Nested dictionaries are printed without error."""
+def test_print_nested_dict_prints_without_error(caplog):
+    """Nested dictionaries are logged without error."""
+    import logging  # noqa: PLC0415
+
     d = {"key1": {"a": 1, "b": 2}, "key2": {"c": 3}}
-    print_nested_dict(d)
-    captured = capsys.readouterr()
-    assert "key1" in captured.out
-    assert "key2" in captured.out
+    with caplog.at_level(logging.WARNING):
+        print_nested_dict(d)
+    assert "key1" in caplog.text
+    assert "key2" in caplog.text
 
 
 def test_populate_useful_dicts_returns_three_dicts():
@@ -2860,36 +2120,6 @@ def test_pivot_table_rounding(data):
     )
     assert isinstance(result, pd.DataFrame)
     assert not result.empty
-
-
-# ---------------------------------------------------------------------------
-# acro_tables.py — surv_func federated mode with output='plot' (lines 650-655)
-# ---------------------------------------------------------------------------
-
-
-def test_surv_func_federated_plot():
-    """Surv_func() in federated mode with output='plot' stores evidence (lines 650-655)."""
-    acro_obj = ACRO(federated=True)
-    np.random.seed(42)
-    mock_df = pd.DataFrame(
-        {
-            "futime": np.random.exponential(100, 100),
-            "death": np.random.binomial(1, 0.3, 100),
-        }
-    )
-    result = acro_obj.surv_func(
-        mock_df.futime, mock_df.death, output="plot", filename="test_km.png"
-    )
-    # Should return a tuple (None, filename)
-    assert result is not None
-    assert isinstance(result, tuple)
-
-
-def test_pie_blocked_extension_custom_test(data):
-    """Pie() with a blocked file extension returns None without running checks."""
-    acro_obj = ACRO(suppress=False)
-    result = acro_obj.pie(data, column="grant_type", filename="chart.svg")
-    assert result is None
 
 
 def test_sdc_checks_unknown_analysis_returns_review() -> None:
